Length statistics after merging

vs_merging_lengths computes length statistics for forward reads, reverse reads, merged reads, and their overlaps before and after merging.

Usage

vs_merging_lengths(
  fastq_input,
  reverse,
  minovlen = 10,
  minlen = 0,
  threads = 1,
  plot_title = TRUE
)

Arguments

fastq_input

A FASTQ file path or FASTQ object containing (forward) reads. See Details.

reverse

A FASTQ file path or FASTQ object containing (reverse) reads.

minovlen

Minimum overlap between the merged reads. Must be at least 5. Defaults to 10.

minlen

Minimum number of bases a sequence must have to be retained. Defaults to 0. See Details.

threads

Number of computational threads to be used by VSEARCH. Defaults to 1.

plot_title

If TRUE (default), a summary title will be displayed in the plot. Set to FALSE for no title.

Value

A tibble with the following columns:

length_1: The length of the forward reads.

length_2: The length of the reverse reads.

length_merged: The length of the merged reads.

length_overlap: The length of the overlap between the forward and reverse reads.

In case of missing values for the latter two columns, it means that the corresponding reads were not merged.

The tibble includes additional attributes:

plot

A ggplot2 object visualizing the returned data frame.

statistics

Additional statistics returned from vs_fastq_mergepairs.

Details

The function uses vs_fastq_mergepairs where the arguments to this function are described in detail.

These length statistics are most typically used in order to tune the filter and trimming of reads such that the merged reads are of high quality.

References

https://github.com/torognes/vsearch

See also

Examples

if (FALSE) { # \dontrun{
# Define arguments
R1.file <- file.path(file.path(path.package("Rsearch"), "extdata"),
                     "small_R1.fq")
R2.file <- file.path(file.path(path.package("Rsearch"), "extdata"),
                     "small_R2.fq")

# Run function
merging.tbl <- vs_merging_lengths(fastq_input = R1.file,
                                  reverse = R2.file)

# Display plot
merging_stats_plot <- attr(merging.tbl, "plot")
print(merging_stats_plot)

} # }